requirements.txt
requests>=2.21.1
pyhamcrest>=1.9.0
allure-pytest==2.8.16